+++ packages/sources/dar/src/endpoint/price.ts
@@ -47,7 +47,7 @@ export const priceTransport = new WebSocketTransport<PriceEndpointTypes>({
               result: message.price,
             },
             timestamps: {
-              providerIndicatedTime: message.effectiveTime * 1000,
+              providerIndicatedTimeUnixMs: message.effectiveTime * 1000,
             },
           },
         },
@@ -60,4 +60,5 @@ export const priceEndpoint = new PriceEndpoint({
   name: 'price',
   transport: priceTransport,
   inputParameters: priceEndpointInputParameters,
+  aliases: ['crypto'],
 })