+++ packages/composites/crypto-volatility-index/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @chainlink/crypto-volatility-index-adapter
 
+## 1.3.35
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.9.1
+
+## 1.3.34
+
+### Patch Changes
+
+- Updated dependencies [6e9f97e33]
+  - @chainlink/token-allocation-adapter@1.9.0
+
 ## 1.3.33
 
 ### Patch Changes