+++ packages/composites/curve-3pool/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @chainlink/curve-3pool-adapter
 
+## 3.0.16
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.9.1
+
+## 3.0.15
+
+### Patch Changes
+
+- Updated dependencies [6e9f97e33]
+  - @chainlink/token-allocation-adapter@1.9.0
+
 ## 3.0.14
 
 ### Patch Changes