+++ packages/composites/outlier-detection/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @chainlink/outlier-detection-adapter
 
+## 1.2.63
+
+### Patch Changes
+
+- @chainlink/ea@1.3.62
+
+## 1.2.62
+
+### Patch Changes
+
+- @chainlink/ea@1.3.61
+
+## 1.2.61
+
+### Patch Changes
+
+- @chainlink/ea@1.3.60
+
+## 1.2.60
+
+### Patch Changes
+
+- @chainlink/ea@1.3.59
+
 ## 1.2.59
 
 ### Patch Changes