+++ packages/sources/ncfx/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @chainlink/ncfx-adapter
 
+## 3.0.2
+
+### Patch Changes
+
+- 3dd345367: Fixed inverses issue. Bumped v3 framework version.
+
+## 3.0.1
+
+### Patch Changes
+
+- 4bade45ad: Fixed crypto websocket expected connection message
+
 ## 3.0.0
 
 ### Major Changes